Moved top level app event handlers into a dedicated class.
