+++ src/lib/components/AppComponent.js
@@ -1,16 +1,12 @@
 const Layers = require('../app/AppLayers.js');
 const { LitElement, html, css } = require('lit-element');
 const { grid } = require('./SharedCss.js');
-const { convertToCell } = require('../ui/CanvasUtilities.js');
-const WorkerCommands = require('../workers/WorkerCommands.js');
 const AppBuilder = require('../app/AppBuilder.js');
+const AppEventHandler = require('./AppEventHandler.js');
 
 const {
 	updateConfiguredLandscape,
 } = require('../ui/UIConfigurationUtilities.js');
-
-const DISPLAY_TRANSITION_ERR_MSG =
-	'There was an error attempting to change display modes.';
 
 /**
  * The top level component for the app.
@@ -19,6 +15,13 @@ class AppComponent extends LitElement {
 	constructor() {
 		super();
 		AppBuilder.buildApp(this);
+		this.handler = new AppEventHandler(
+			this,
+			this.shadowRoot,
+			this.config,
+			this.stateManager,
+			this.displayManager
+		);
 	}
 
 	/**
@@ -89,36 +92,48 @@ class AppComponent extends LitElement {
 					<div class="container row">
 						<game-selector
 							event="game-changed"
-							@game-changed=${this.handleGameChanged}
+							@game-changed=${this.handler.gameChanged.bind(this.handler)}
 						></game-selector>
 						<cell-size-control
 							event="cell-size-changed"
 							min="5"
 							max="100"
 							value="20"
-							@cell-size-changed=${this.changedCellSize}
+							@cell-size-changed=${this.handler.cellSizeChanged.bind(
+								this.handler
+							)}
 						></cell-size-control>
 						<speed-selector
 							event="speed-changed"
-							@speed-changed=${this.handleGameSpeedChanged}
+							@speed-changed=${this.handler.gameSpeedChanged.bind(this.handler)}
 						></speed-selector>
 						<start-button
 							state="IDLE"
-							@sim-event-start-requested=${this.handleStartButtonClicked}
-							@sim-event-pause-requested=${this.handlePauseButtonClicked}
-							@sim-event-resume-requested=${this.handleResumeButtonClicked}
+							@sim-event-start-requested=${this.handler.startButtonClicked.bind(
+								this.handler
+							)}
+							@sim-event-pause-requested=${this.handler.pauseButtonClicked.bind(
+								this.handler
+							)}
+							@sim-event-resume-requested=${this.handler.resumeButtonClicked.bind(
+								this.handler
+							)}
 						></start-button>
 						<event-button
 							id="reset_button"
 							event="sim-reset-requested"
-							@sim-reset-requested=${this.resetSimulation}
+							@sim-reset-requested=${this.handler.resetSimulation.bind(
+								this.handler
+							)}
 						>
 							Reset
 						</event-button>
 						<event-button
 							id="fullscreen_button"
 							event="fullscreen-requested"
-							@fullscreen-requested=${this.launchFullScreen}
+							@fullscreen-requested=${this.handler.launchFullScreen.bind(
+								this.handler
+							)}
 						>
 							Fullscreen
 						</event-button>
@@ -127,7 +142,9 @@ class AppComponent extends LitElement {
 						<event-checkbox
 							id="display_grid"
 							event="dispay-grid-toggle"
-							@dispay-grid-toggle=${this.handleGridBackgroundClicked}
+							@dispay-grid-toggle=${this.handler.gridBackgroundClicked.bind(
+								this.handler
+							)}
 						>
 							Display Grid
 						</event-checkbox>
@@ -135,7 +152,9 @@ class AppComponent extends LitElement {
 						<event-checkbox
 							id="display_fullscreen"
 							event="enable-fullscreen-toggle"
-							@enable-fullscreen-toggle=${this.handleFullScreenClicked}
+							@enable-fullscreen-toggle=${this.handler.fullScreenClicked.bind(
+								this.handler
+							)}
 						>
 							Fullscreen
 						</event-checkbox>
@@ -143,23 +162,30 @@ class AppComponent extends LitElement {
 						<event-checkbox
 							id="random_start"
 							event="random-start-toggle"
-							@random-start-toggle=${this.handleRandomStartClicked}
+							@random-start-toggle=${this.handler.randomStartClicked.bind(
+								this.handler
+							)}
 						>
 							Random Start
 						</event-checkbox>
 					</div>
 				</div>
-				<div id="canvas_container" @contextmenu="${this.displayContextMenu}">
+				<div
+					id="canvas_container"
+					@contextmenu="${this.handler.displayContextMenu.bind(this.handler)}"
+				>
 					<canvas id="grid_canvas"></canvas>
 					<canvas id="sim_canvas"></canvas>
 					<canvas
 						id="draw_canvas"
-						@click=${this.handleDrawCanvasClicked}
-						@mousemove=${this.handleDrawCanvasMouseMoved}
+						@click=${this.handler.drawCanvasClicked.bind(this.handler)}
+						@mousemove=${this.handler.drawCanvasMouseMoved.bind(this.handler)}
 					></canvas>
 					<context-menu
 						event="context-menu-command"
-						@context-menu-command=${this.handleContextMenuCommand}
+						@context-menu-command=${this.handler.contextMenuCommand.bind(
+							this.handler
+						)}
 					></context-menu>
 				</div>
 				<div id="status_bar">
@@ -174,14 +200,6 @@ class AppComponent extends LitElement {
 				</div>
 			</div>
 		`;
-	}
-
-	handleGameSpeedChanged(event) {
-		this.config.game.tickLength = event.detail.fps.tickLength;
-		this.stateManager.sendWorkerMessage(Layers.SIM, {
-			command: WorkerCommands.LifeSystemCommands.SET_CONFIG,
-			config: this.config,
-		});
 	}
 
 	/**
@@ -224,24 +242,6 @@ class AppComponent extends LitElement {
 		return document.fullscreenElement != null;
 	}
 
-	calculateConfiguredCanvasHeight() {
-		// Note: This will use the same padding/margins as the HTML Body.
-		let blockElement = this.shadowRoot.getElementById('block');
-		let headerElement = this.shadowRoot.getElementById('header');
-		let controlBarElement = this.shadowRoot.getElementById('control_bar');
-		let statusBarElement = this.shadowRoot.getElementById('status_bar');
-		let bodyMargin = 8 * 2; //Padding on body element in CSS is 8 top and bottom.
-
-		return (
-			window.innerHeight -
-			bodyMargin -
-			(blockElement.offsetHeight +
-				headerElement.offsetHeight +
-				controlBarElement.offsetHeight +
-				statusBarElement.offsetHeight)
-		);
-	}
-
 	/**
 	 * Event handler for reacting to when the hosting web page is resized.
 	 * @param {*} event
@@ -250,30 +250,8 @@ class AppComponent extends LitElement {
 	handlePageResize(event) {
 		this.sizeCanvas();
 		updateConfiguredLandscape(this.config);
-		this.refreshGrid();
-		this.setflagAsDirty(Layers.DRAWING);
-	}
-
-	refreshGrid() {
-		this.stateManager.displayGrid
-			? this.requestToDrawGrid()
-			: this.stateManager.clearRender(Layers.GRID);
-	}
-
-	/**
-	 * Requests the grid worker to generate a grid scene.
-	 * @private
-	 */
-	requestToDrawGrid() {
-		this.stateManager.sendWorkerMessage(Layers.GRID, {
-			command: WorkerCommands.LifeCycle.PROCESS_CYCLE,
-			parameters: {
-				cellWidth: this.config.zoom,
-				cellHeight: this.config.zoom,
-				gridWidth: this.config.canvas.width,
-				gridHeight: this.config.canvas.height,
-			},
-		});
+		this.handler.refreshGrid();
+		this.handler.setflagAsDirty(Layers.DRAWING);
 	}
 
 	/**
@@ -283,10 +261,13 @@ class AppComponent extends LitElement {
 	 * @returns {App} The instance.
 	 */
 	updateUI(message) {
-		message.aliveCellsCount && this.setAliveCellsCount(message.aliveCellsCount);
+		message.aliveCellsCount &&
+			this.handler.setAliveCellsCount(message.aliveCellsCount);
 
 		message.numberOfSimulationIterations &&
-			this.setSimGenerationCountComponent(message.numberOfSimulationIterations);
+			this.handler.setSimGenerationCountComponent(
+				message.numberOfSimulationIterations
+			);
 
 		if (message.origin && message.origin == Layers.DRAWING && message.stack) {
 			this.manageStartButtonEnablement(
@@ -319,47 +300,6 @@ class AppComponent extends LitElement {
 		}
 	}
 
-	/**
-	 * Resets all web workers and the UI.
-	 */
-	resetSimulation() {
-		this.stateManager.stopSimulation();
-		this.shadowRoot.querySelector(
-			'context-menu'
-		).updateCommandState = JSON.stringify({
-			key: 'runSim',
-			activeState: 'start',
-		});
-		this.transitionToTheStartButton()
-			.setAliveCellsCount(0)
-			.setSimGenerationCountComponent(0);
-		return this.stateManager.allowDrawing().resetSimulation();
-	}
-
-	/**
-	 * Changes the current state of the simulation button.
-	 * @private
-	 * @returns {Main} Returns the instance of the main thread being modified.
-	 */
-	transitionToTheStartButton() {
-		this.shadowRoot.querySelector('start-button').state = 'IDLE';
-		return this;
-	}
-
-	setAliveCellsCount(count) {
-		this.shadowRoot
-			.getElementById('alive_cells_count')
-			.setAttribute('value', count);
-		return this;
-	}
-
-	setSimGenerationCountComponent(count) {
-		this.shadowRoot
-			.getElementById('sim_generation_count')
-			.setAttribute('value', count);
-		return this;
-	}
-
 	getCanvasContext(elementId) {
 		let canvas = this.shadowRoot.getElementById(elementId);
 		if (!canvas) {
@@ -368,226 +308,8 @@ class AppComponent extends LitElement {
 		return canvas.getContext('2d');
 	}
 
-	/**
-	 * Event handler for processing a user click when in drawing mode.
-	 * @param {Event} clickEvent Event generated when the draw canvas is clicked.
-	 */
-	handleDrawCanvasClicked(clickEvent) {
-		let menu = this.shadowRoot.querySelector('context-menu');
-		if (menu.display) {
-			menu.display = false;
-			return;
-		}
-
-		if (this.stateManager.isDrawingAllowed()) {
-			let boundary = this.shadowRoot
-				.getElementById('draw_canvas')
-				.getBoundingClientRect();
-			let cellLocation = convertToCell(clickEvent, boundary, this.config.zoom);
-			this.setflagAsDirty(Layers.DRAWING);
-			this.stateManager.sendWorkerMessage(Layers.DRAWING, {
-				command: WorkerCommands.DrawingSystemCommands.TOGGLE_CELL,
-				cx: cellLocation.x,
-				cy: cellLocation.y,
-			});
-		}
-	}
-
-	displayContextMenu(clickEvent) {
-		clickEvent.preventDefault();
-		let boundary = this.shadowRoot
-			.getElementById('draw_canvas')
-			.getBoundingClientRect();
-		let contextMenu = this.shadowRoot.querySelector('context-menu');
-
-		contextMenu.menuPosition = {
-			clickEvent: clickEvent,
-			boundary: boundary,
-			zoom: this.config.zoom,
-		};
-
-		contextMenu.display = true;
-		return false;
-	}
-
-	handleDrawCanvasMouseMoved(event) {
-		let boundary = this.shadowRoot
-			.getElementById('draw_canvas')
-			.getBoundingClientRect();
-		let cellLocation = convertToCell(event, boundary, this.config.zoom);
-		this.stateManager.setActiveCell(cellLocation);
-	}
-
-	handleStartButtonClicked() {
-		this.shadowRoot.querySelector(
-			'context-menu'
-		).updateCommandState = JSON.stringify({
-			key: 'runSim',
-			activeState: 'pause',
-		});
-		return new Promise((resolve, reject) => {
-			Promise.resolve(
-				this.displayManager.setDisplayMode(
-					this.stateManager.getDisplayPreference()
-				)
-			)
-				.catch((reason) => {
-					console.error(DISPLAY_TRANSITION_ERR_MSG);
-					console.error(reason);
-				})
-				.then(() => {
-					document.fullscreenElement && this.handlePageResize();
-					this.stateManager.preventDrawing();
-					this.stateManager.startSimulation();
-				});
-		});
-	}
-
-	handlePauseButtonClicked() {
-		this.stateManager.stopSimulation();
-		this.stateManager.pauseSimulationInDrawingMode();
-		this.shadowRoot.querySelector(
-			'context-menu'
-		).updateCommandState = JSON.stringify({
-			key: 'runSim',
-			activeState: 'resume',
-		});
-		this.setflagAsDirty(Layers.DRAWING);
-	}
-
-	handleResumeButtonClicked() {
-		this.shadowRoot.querySelector(
-			'context-menu'
-		).updateCommandState = JSON.stringify({
-			key: 'runSim',
-			activeState: 'pause',
-		});
-		Promise.resolve(
-			this.displayManager.setDisplayMode(
-				this.stateManager.getDisplayPreference()
-			)
-		)
-			.catch((reason) => {
-				console.error(DISPLAY_TRANSITION_ERR_MSG);
-				console.error(reason);
-			})
-			.then(() => {
-				document.fullscreenElement && this.handlePageResize();
-				this.stateManager.preventDrawing();
-				this.stateManager.startSimulation();
-			});
-	}
-
-	launchFullScreen() {
-		return new Promise((resolve, reject) => {
-			let container = this.shadowRoot.getElementById('canvas_container');
-			Promise.resolve(this.displayManager.setDisplayMode(true, container))
-				.catch((reason) => {
-					console.error(DISPLAY_TRANSITION_ERR_MSG);
-					console.error(reason);
-					reject();
-				})
-				.then(() => {
-					document.fullscreenElement && this.handlePageResize();
-					resolve();
-				});
-		});
-	}
-
 	setCellShapeOption(event) {
 		this.config.cell.shape = event.detail.shape;
-	}
-
-	/**
-	 * Command all registered workers to set their cell size.
-	 */
-	changedCellSize(event) {
-		this.config.zoom = event.detail.cellSize;
-		updateConfiguredLandscape(this.config);
-		//Inform the drawing system and Life Simulation of the change.
-		this.stateManager.broadcast({
-			command: WorkerCommands.LifeSystemCommands.SET_CELL_SIZE,
-			cellSize: this.config.zoom,
-		});
-		this.refreshGrid();
-		this.setflagAsDirty(Layers.DRAWING);
-	}
-
-	/**
-	 * Event handler for when the grid checkbox is clicked.
-	 */
-	handleGridBackgroundClicked(event) {
-		this.stateManager.displayGrid = event.detail.checked;
-		this.refreshGrid();
-	}
-
-	handleFullScreenClicked(event) {
-		this.stateManager.setDisplayPreference(event.detail.checked);
-	}
-
-	handleRandomStartClicked(event) {
-		this.stateManager.setRandomStartPreference(event.detail.checked);
-	}
-
-	/**
-	 * Handles processing the context menu item clicked.
-	 * @param {number} row - The horizontal coordinate of the cell clicked.
-	 * @param {number} col - The vertical coordinate of the cell clicked.
-	 * @param {string} cmdName - The command clicked in the context menu
-	 */
-	handleContextMenuCommand(event) {
-		event.detail.simCommand
-			? this.processContextMenuSimCommand(event)
-			: this.generateTemplate(event);
-	}
-
-	generateTemplate(event) {
-		this.config.elementaryCAs.useRandomStart = this.stateManager.getRandomStartPreference();
-		this.stateManager.sendWorkerMessage(Layers.DRAWING, {
-			command: WorkerCommands.DrawingSystemCommands.DRAW_TEMPLATE,
-			templateName: event.detail.command,
-			row: event.detail.row,
-			col: event.detail.col,
-			config: this.config,
-		});
-		this.setflagAsDirty(Layers.DRAWING);
-	}
-
-	setflagAsDirty(workerName) {
-		this.stateManager.workerSystem.setWorkerDirtyFlag(workerName, true);
-		return this;
-	}
-
-	processContextMenuSimCommand(event) {
-		let startButton = this.shadowRoot.querySelector('start-button');
-		switch (event.detail.command) {
-			case 'start-sim':
-				startButton.state = 'RUNNING';
-				this.handleStartButtonClicked();
-				break;
-			case 'pause-sim':
-				startButton.state = 'PAUSED';
-				this.handlePauseButtonClicked();
-				break;
-			case 'resume-sim':
-				startButton.state = 'RUNNING';
-				this.handleResumeButtonClicked();
-				break;
-			case 'reset':
-				this.resetSimulation();
-				break;
-			default:
-				throw new Error('Unknown context menu command.');
-		}
-		return;
-	}
-
-	handleGameChanged(event) {
-		this.config.game.activeGame = event.detail.game;
-		this.stateManager.sendWorkerMessage(Layers.SIM, {
-			command: WorkerCommands.LifeSystemCommands.SET_CONFIG,
-			config: this.config,
-		});
 	}
 }
